+++ packages/sodium/lib/src/ffi/api/crypto_ffi.dart
@@ -13,7 +13,8 @@ import '../../api/short_hash.dart';
 import '../../api/sign.dart';
 import '../../api/sumo/pwhash.dart';
 import '../bindings/libsodium.ffi.dart';
-import 'aead_ffi.dart';
+import 'aead_chacha20poly1305_ffi.dart';
+import 'aead_xchacha20poly1305ietf_ffi.dart';
 import 'auth_ffi.dart';
 import 'box_ffi.dart';
 import 'generic_hash_ffi.dart';
@@ -41,7 +42,14 @@ class CryptoFFI implements Crypto {
   late final SecretStream secretStream = SecretStreamFFI(sodium);
 
   @override
-  late final Aead aead = AeadFFI(sodium);
+  late final Aead aead = aeadXChaCha20Poly1305IETF;
+
+  @override
+  late final Aead aeadChaCha20Poly1305 = AeadChacha20Poly1305FFI(sodium);
+
+  @override
+  late final Aead aeadXChaCha20Poly1305IETF =
+      AeadXChaCha20Poly1305IETFFFI(sodium);
 
   @override
   late final Auth auth = AuthFFI(sodium);